requirements.txt
google-genai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
